+++ R/import_gtfs.R
@@ -85,12 +85,10 @@ import_gtfs <- function(path,
   assert_vector(skip, "character", null_ok = TRUE)
   assert_vector(encoding, "character", len = 1L, subset_of = val_enc)
 
-  if (!grepl("\\.zip$", path)) error_path_must_be_zip()
-
   path_is_url <- grepl("^http[s]?\\:\\/\\/\\.*", path)
 
-  if (!path_is_url & !file.exists(path)) error_non_existent_file(path)
-  if (!is.null(files) & !is.null(skip)) error_files_and_skip_provided()
+  if (!path_is_url && !file.exists(path)) error_non_existent_file(path)
+  if (!is.null(files) && !is.null(skip)) error_files_and_skip_provided()
 
   for (input_types in extra_spec) {
     if (any(! input_types %chin% c("character", "integer", "numeric"))) {
@@ -101,19 +99,35 @@ import_gtfs <- function(path,
   # if 'path' is an URL, download it and save path to downloaded file to 'path'
 
   if (path_is_url) {
-
     tmp <- tempfile(pattern = "gtfs", fileext = ".zip")
     utils::download.file(path, tmp, method = "auto", quiet = quiet)
 
     if (!quiet) message("File downloaded to ", tmp, ".")
 
     path <- tmp
-
   }
 
-  # retrieve which files are inside the GTFS and remove '.txt' from their names
+  # check which files are inside the GTFS. if any non text file is found, raise
+  # a warning and do not try to read it as a csv. remove the '.txt' extension
+  # from the text files to reference them without it in messages and errors
 
-  files_in_gtfs <- zip::zip_list(path)$filename
+  files_in_gtfs <- tryCatch(
+    zip::zip_list(path)$filename,
+    error = function(cnd) cnd
+  )
+  if (inherits(files_in_gtfs, "error")) error_path_must_be_zip()
+
+  non_text_files <- files_in_gtfs[!grepl("\\.txt$", files_in_gtfs)]
+
+  if (!identical(non_text_files, character(0))) {
+    warning(
+      "Found non .txt files when attempting to read the GTFS feed: ",
+      paste(non_text_files, collapse = ", "), "\n",
+      "These files have been ignored and were not imported to the GTFS object.",
+      call. = FALSE
+    )
+  }
+  files_in_gtfs <- setdiff(files_in_gtfs, non_text_files)
   files_in_gtfs <- gsub("\\.txt", "", files_in_gtfs)
 
   # read only the text files specified either in 'files' or in 'skip'.
@@ -258,25 +272,28 @@ read_files <- function(file,
   }
 
   # read 'file' first row to figure out which fields are present
-  # if 'file_standards' is NULL then file is undocumented
+  # - if 'file_standards' is NULL then file is undocumented
+  # - print warning message if warning is raised and 'quiet' is FALSE
 
   if (is.null(file_standards) & !quiet) {
     message("  - File undocumented. Trying to read it as a csv.")
   }
 
-  sample_dt <- data.table::fread(
-    file.path(tmpdir, file_txt),
-    nrows = 1,
-    colClasses = "character"
+  withCallingHandlers(
+    {
+      sample_dt <- data.table::fread(
+        file.path(tmpdir, file_txt),
+        nrows = 1,
+        colClasses = "character"
+      )
+    },
+    warning = function(cnd) if (!quiet) message("  - ", conditionMessage(cnd))
   )
 
   # if 'file' is completely empty (even without a header), return a NULL
   # 'data.table'
 
-  if (ncol(sample_dt) == 0) {
-    if (!quiet) message("  - File is empty. Returning a NULL data.table")
-    return(data.table::data.table(NULL))
-  }
+  if (ncol(sample_dt) == 0) return(data.table::data.table(NULL))
 
   # retrieve which fields are inside the file
 
@@ -332,7 +349,7 @@ read_files <- function(file,
   fields_classes <- c(doc_classes, undoc_classes)
 
   # read the file specifying the column classes
-  # if a warning is thrown (e.g. due to a parsing failure) and 'quiet' != FALSE,
+  # if a warning is thrown (e.g. due to a parsing failure) and 'quiet' is FALSE,
   # print warning message to console to help debugging (otherwise all warnings
   # messages are thrown simultaneously at the end, which doesn't help as much)
 
@@ -356,7 +373,7 @@ read_files <- function(file,
 
 
 error_path_must_be_zip <- parent_function_error(
-  "'path' must have '.zip' extension.",
+  "Could not unzip file. Please make sure 'path' points to a zip file/url.",
   subclass = "path_must_be_zip"
 )